fix(node-gyp): install python for node-gyp
Don't install node-gyp directly as newer versions of node bundle this and use that version even if
you have another one installed. node-gyp@8, which is the version bundled with the latest npm (v8)
requires Python 3.6 or newer. In stretch, the python3 package is 3.5. To allow node-gyp to work,
install npm@7 which works with Python 2.7. In Buster continue to install the latest and install
Python3.7. In Lambda images don't install node-gyp and install npm@7. Also bump the Yarn version to
the latest classic version