feat: add retry script and attempt refactor
This commit adds a retry script to `/usr/local/bin` for dependent images
to make use of if they want to. For example, `yarn install
--frozen-lockfile` might return 503s periodically, and someone may want
to retry that command a few times before giving up.

This commit also creates `scripts/` and `policies/` folders and has each
image pull from those instead of duplicating the same policies and
scripts over and over. I am unsure if this is wise or will even work,
but it is an attempt.