Attempt separate CI & CD scripts
This commit moves `build.yml` over to `cd.yml` and adds a `ci.yml` to
run on pull requests only. It sets `strategy.fail-fast: true`,
build-push-action's `push: false`, and sets up a cache to use a scoped
`gha` cache. If this was done correctly, the goal is to be able to test
the build on PRs and not publish.