chore: run builds on PRs
This updates the `build.yml` file to run on PRs and pushes to them while
also attempting to only publish when we are in the context of the `main`
branch.

This also updates the `fail-fast` to stop builds if there is anything
wrong with any of them to cut down on build times if you're going to fix
the problem and push again, triggering another build.