feat: add 18-lambda image
Lambda now supports Node 18.

Remove yarn since there are different ways to install depending on the
version. If the project is using yarn, install the version they need.

Remove the aws-sdk npm package. v3 created different packages for each
service. The runtime should also provide the sdk, so we don't need that
installed.

Remove the service user. This image should used for building Lambda
packages, so we don't care what user the container is using.

Remove the service directory. If we want to test the Lambda locally, it
assumes it's ran in /var/tasks. Our pipeline doesn't care where the
working directory is.

Cleanup README. Call out the recommended images.